# openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: RESTful EPP
  version: "1.0"
servers:
- url: http://localhost:8080/repp/v1/
paths:
...
  /hosts:
    post:
      operationId: hosts_post
      parameters:
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Host create
      tags:
      - hosts
      x-openapi-router-controller: openapi_server.controllers.hosts_controller
  /hosts/{id}:
    delete:
      operationId: hosts_id_delete
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Host delete
      tags:
      - hosts
      x-openapi-router-controller: openapi_server.controllers.hosts_controller
    get:
      operationId: hosts_id_get
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: filter
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: val
        required: false
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Host info
      tags:
      - hosts
      x-openapi-router-controller: openapi_server.controllers.hosts_controller
    head:
      operationId: hosts_id_head
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          description: OK
          headers:
            REPP-Check-Avail:
              $ref: '#/components/headers/check-avail'
            REPP-Check-Reason:
              $ref: '#/components/headers/check-reason'
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Host check
      tags:
      - hosts
      x-openapi-router-controller: openapi_server.controllers.hosts_controller
    patch:
      operationId: hosts_id_patch
      parameters:
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Host update
      tags:
      - hosts
      x-openapi-router-controller: openapi_server.controllers.hosts_controller
...
components:
  headers:
    check-avail:
      description: Result of object check
      explode: false
      required: true
      schema:
        type: integer
      style: simple
    check-reason:
      description: Explanation why object cannot be provisioned
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    cltrid:
      description: Client transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
    eppcode:
      description: EPP result code
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    queue-size:
      description: Number of messages in the queue
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    svtrid:
      description: Server transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
...
  requestBodies:
    defaultOptionalRequestBody:
      content:
        application/epp+json:
          schema:
            type: string
        application/epp+xml:
          schema:
            type: string
      description: Default request body for XML of JSON message
      required: false
    defaultRequestBody:
      content:
        application/epp+json:
          schema:
            type: string
        application/epp+xml:
          schema:
            type: string
      description: Default request body for XML of JSON message
      required: true
...